+++ asapdiscovery-data/asapdiscovery/data/schema_v2/target.py
@@ -1,6 +1,9 @@
 from pathlib import Path
 from typing import Any, Dict, Optional, Tuple, Union  # noqa: F401
 
+from asapdicovery.data.schema_v2.identifiers import TargetIdentifiers
+from pydantic import Field, root_validator
+
 from asapdiscovery.data.openeye import (
     bytes64_to_oedu,
     load_openeye_design_unit,
@@ -11,10 +14,8 @@ from asapdiscovery.data.openeye import (
     pdb_string_to_oemol,
     save_openeye_design_unit,
 )
-from asapdiscovery.data.postera.manifold_data_validation import TargetTags
 from asapdiscovery.modeling.modeling import split_openeye_mol
 from asapdiscovery.modeling.schema import MoleculeFilter
-from pydantic import Field, root_validator
 
 from .schema_base import (
     DataModelAbstractBase,
@@ -27,25 +28,6 @@ from .schema_base import (
 
 class InvalidTargetError(ValueError):
     ...
-
-
-class TargetIdentifiers(DataModelAbstractBase):
-    """
-    Identifiers for a Target
-    """
-
-    target_type: Optional[TargetTags] = Field(
-        None,
-        description="Tag describing the target type e.g SARS-CoV-2-Mpro, etc.",
-    )
-
-    fragalysis_id: Optional[str] = Field(
-        None, description="The Fragalysis ID of the target if applicable"
-    )
-
-    pdb_code: Optional[str] = Field(
-        None, description="The PDB code of the target if applicable"
-    )
 
 
 class Target(DataModelAbstractBase):
